requirements.txt
arxiv>=2.1.3
openai>=1.60.0
json-repair>=0.30.0
jinja2>=3.1.4
requests>=2.32.0
pyyaml>=6.0.2
numpy>=1.26.0
python-dotenv>=1.0.0
flask>=3.1.0
PyNaCl>=1.5.0
